requirements.txt
python-telegram-bot>=13.4
APScheduler>=3.6
croniter>=1.0